feat: CB batching + wave32 fix + bench TDR guard + shim_plan
- runtime/buffer.cpp: command-buffer batching infrastructure
  vkflame_batch_begin/end collapses 3 vkQueueSubmit calls per _staged()
  op into 1, reducing WDDM overhead from ~9ms to ~3ms per kernel call
  Includes TRANSFER->COMPUTE barrier injection and deferred pool destroy

- runtime/dispatch.cpp: batch-aware do_dispatch
  Reuses batch CB when active; inserts pipeline barriers between stages
  Falls back to per-op allocate+submit when batch mode is off

- runtime/pipeline.cpp: enforce wave32 for rms_norm and softmax kernels
  VK_PIPELINE_SHADER_STAGE_REQUIRED_SUBGROUP_SIZE_CREATE_INFO = 32
  Fixes silent misbehaviour on RDNA4 where driver reports subgroupSize=64

- runtime/device.cpp: report minSubgroupSize (=32) not subgroupSize (=64)
  RDNA4 driver lies; subgroupSizeControl.minSubgroupSize reveals truth

- kernels/linear_fp16_fp32out.glsl: fix output type (float -> float16_t)
  Accumulator stays float (fp32 internal), output write is float16_t
  Eliminates out-of-bounds write that was corrupting fp16 output buffers

- torch_backend/ops.py: batch mode wired into _staged context manager
  Detects vkflame_batch_begin symbol; routes through batch path when present

- tools/gpu_bench.py: per-iter timing + 3s deadline cap (TDR guard)
  Median instead of mean; _N_BENCH=20; early exit if deadline exceeded

- shim_plan.md: documented active defects (P0 TDR, P1 mm precision,
  P2 rms_norm precision) with exact fix paths and performance targets

- .gitignore: exclude bench_out.txt / bench_stderr.txt scratch files

+++ kernels/linear_fp16_fp32out.glsl
@@ -9,8 +9,8 @@ layout(local_size_x = 16, local_size_y = 16) in;
 
 layout(set = 0, binding = 0) readonly  buffer BufA { float16_t a[]; };
 layout(set = 0, binding = 1) readonly  buffer BufB { float16_t b[]; };
-layout(set = 0, binding = 2) readonly  buffer BufC { float     c[]; };  // fp32 bias (unused)
-layout(set = 0, binding = 3) writeonly buffer BufD { float     d[]; };  // fp32 output
+layout(set = 0, binding = 2) readonly  buffer BufC { float16_t c[]; };  // bias (unused)
+layout(set = 0, binding = 3) writeonly buffer BufD { float16_t d[]; };  // fp16 output
 
 layout(push_constant) uniform PC {
     uint M;
@@ -46,5 +46,5 @@ void main() {
     for (uint k = 0u; k < pc.K; k++)
         acc += float(a[row * pc.K + k]) * float(b[k * pc.N + col]);
 
-    d[row * pc.N + col] = apply_act(acc);  // write float32, not float16_t
+    d[row * pc.N + col] = float16_t(apply_act(acc));  // accumulate fp32, write fp16
 }
